cleanup main and move sensor code to sensors.h
